% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichState.R
\name{whichState}
\alias{whichState}
\title{whichState}
\usage{
whichState(phoneNumber, usa = TRUE)
}
\arguments{
\item{phoneNumber}{US Phone Number: A string of 10 digits. There should be no spacing or punctuations between digits}

\item{usa}{Set to TRUE by default. If usa is not true, the function cannot be executed because the US State Area codes only work for US Phone Numbers}
}
\value{
The US State where the Phone Number is from
}
\description{
Find out which American State is the phone number from via US Area Code.
}
\details{
This function has been designed to extract the first 3 digits of the provided phone number, match them to the US State Area Codes and to find out which state the phone number is from. The phone number provided has to be a proper US phone number with 10 digits.
}
\examples{
whichState(1234567890, usa=TRUE)
}
\author{
Gerald Lee
}
\keyword{USA,}
\keyword{area,}
\keyword{code}
\keyword{numbers,}
\keyword{phone,}
\keyword{state,}
